refactor: minor improvements
